especialidades terminado

+++ 02-medico/src/main/java/com/hospital/medico/Repository/MedicoRepository.java
@@ -3,7 +3,7 @@ import com.hospital.medico.Model.MedicoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-public interface MedicoRepository extends JpaRepository<MedicoModel,Integer> {
+public interface MedicoRepository extends JpaRepository<MedicoModel,Long> {
 
     // encontrar medico por el nombre
     List<MedicoModel> findByNombreMedico(String nombreMedico);
